requirements.txt
Flask==0.10.1
Flask-Cache==0.13.1
Flask-Compress==1.3.2
Flask-DebugToolbar==0.10.0
flask-error==0.1
Flask-ErrorMail==0.2.2
Flask-HTTPAuth==3.2.1
Flask-Mail==0.9.1
Flask-Runner==2.1.1
Flask-Session==0.3.0
hgvs==1.0.0a1
mygene==3.0.0
myvariant==0.3.1
neo4j-driver==1.4.0
pandas==0.19.1
plotly==1.12.12
primer3-py==0.5.1 		# Can't be installed on Windows.
pycrypto==2.6.1
pygr==0.8.2
pysam==0.9.1.4 			# Can't be installed on Windows.
python-bencode==1.0.2
lxml==3.7.3
bs4==0.0.1
passlib==1.7.1
argon2_cffi==16.3.0
regex==2017.6.23 
flask_jwt_extended


